auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "fastapi-utils>=0.8.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "typing-inspect>=0.9.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]
